Mark 1

+++ app/src/main/java/com/example/merchantapp/adapters/FoodMenuAdapter.kt
@@ -1,23 +1,22 @@
-package com.example.merchantapp.classes
+package com.example.merchantapp.adapters
 
+import android.app.AlertDialog
 import android.content.Context
+import android.content.DialogInterface
 import android.content.Intent
+import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.Button
-import android.widget.ImageView
+import android.widget.ImageButton
 import android.widget.LinearLayout
 import android.widget.TextView
-import androidx.core.content.ContextCompat
 import androidx.core.content.ContextCompat.startActivity
 import androidx.recyclerview.widget.RecyclerView
 import com.example.ajubamerchant.classes.AdapterInterface
 import com.example.ajubamerchant.classes.FoodMenu
 import com.example.ajubamerchant.classes.Image
-import com.example.ajubamerchant.classes.Order
 import com.example.merchantapp.FoodListActivity
-import com.example.merchantapp.OrderPanelActivity
 import com.example.merchantapp.R
 
 
@@ -27,17 +26,19 @@ class FoodMenuAdapter(
         RecyclerView.Adapter<FoodMenuAdapter.ViewHolder>(){
 
 
+
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder {
         val itemView = LayoutInflater.from(parent.context)
                 .inflate(R.layout.ticket_food_menu, parent, false)
+
         return ViewHolder(itemView)
     }
     inner class ViewHolder(view: View): RecyclerView.ViewHolder(view) {
 
-        var name: TextView =view.findViewById(R.id.tvMenuTitle)
+        var name: TextView =view.findViewById(R.id.textViewFoodListName)
 
         var layout: LinearLayout =view.findViewById(R.id.llFoodMenu)
-        var delete: Button = view.findViewById(R.id.btMenuDelete)
+        var delete: ImageButton = view.findViewById(R.id.buttonFoodListDelete)
 
 
 
@@ -46,10 +47,21 @@ class FoodMenuAdapter(
     override fun onBindViewHolder(holder: ViewHolder, position: Int) {
         holder.name.text=list[position].category
         holder.delete.setOnClickListener{
+            val dialog= AlertDialog.Builder(context)
+            dialog.setTitle("Confirm delete?")
+            dialog.setPositiveButton("Yes"){ dialogInterface: DialogInterface, i: Int ->
+                adapterInterface.deleteMenu(list[position].category)
+                list.removeAt(position)
+                notifyDataSetChanged()
+            }
+            dialog.setNegativeButton("No"){ dialogInterface: DialogInterface, i: Int ->
 
-            adapterInterface.deleteMenu(list[position].category)
-            notifyDataSetChanged()
+            }
+            dialog.create().show()
+
+
         }
+        Log.d("adapter",list.toString())
         holder.layout.setOnClickListener{
             val intent =Intent(context,FoodListActivity::class.java)
             intent.putExtra("category",list[position].category)
